feat: convert Form > Masked Input page to TS
